SUP 2961 & SUP 2964 install-docker-compose & install-dockerize commands
